TVec template fixes

+++ libs/JSystem/include/JSystem/JParticle/JPAEmitter.h
@@ -152,8 +152,11 @@ public:
     void setSpread(f32 i_spread) { mSpread = i_spread; }
     void setLocalTranslation(const JGeometry::TVec3<f32>& i_trans) { mLocalTrs.set(i_trans); }
     void setLocalRotation(const JGeometry::TVec3<s16>& i_rot) {
-        mLocalRot.set(i_rot.x * (360.0f / 0xffff), i_rot.y * (360.0f / 0xffff),
-                      i_rot.z * (360.0f / 0xffff));
+        mLocalRot.set(
+            (s16)(i_rot.x * (360.0f / 0xffff)),
+            (s16)(i_rot.y * (360.0f / 0xffff)),
+            (s16)(i_rot.z * (360.0f / 0xffff))
+        );
     }
     void setRateStep(u8 i_step) { mRateStep = i_step; }
